+++ packages/python-ta/CHANGELOG.md
@@ -24,7 +24,7 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Updated `uv.lock` file with latest packages
 - Switched dependabot configuration from `pip` to `uv` to ensure `uv.lock` file is updated
 - Fixed watch tests on Windows by using `threading` rather than `select`
-- Added tests for `static_type_checker.py` to improve coverage for the `_add_message` method
+- Added tests using unsupported Mypy error codes in `StaticTypeChecker` to improve coverage for the `_add_message` method
 
 ## [2.13.1] - 2026-08-12